requirements.txt
Flask 
openai
chromadb
python-dotenv
requests
PyPDF2






